+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/next/FileSystemNextModule.kt
@@ -1,12 +1,15 @@
 package expo.modules.filesystem.next
 
-import android.net.Uri
+import android.webkit.URLUtil
+import expo.modules.kotlin.Promise
 import expo.modules.kotlin.apifeatures.EitherType
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
 import expo.modules.kotlin.typedarray.TypedArray
 import expo.modules.kotlin.types.Either
 import java.io.File
+import java.io.FileOutputStream
+import java.net.HttpURLConnection
 import java.net.URI
 
 class FileSystemNextModule : Module() {
@@ -14,6 +17,33 @@ class FileSystemNextModule : Module() {
   @OptIn(EitherType::class)
   override fun definition() = ModuleDefinition {
     Name("FileSystemNext")
+
+    AsyncFunction("download") { url: URI, to: FileSystemPath, promise: Promise ->
+      try {
+        with(url.toURL().openConnection() as HttpURLConnection) {
+          // check if code is 200 to 299
+          if (responseCode !in 200..299) {
+            promise.reject(UnableToDownloadException("response has status: $responseCode"))
+            return@AsyncFunction
+          }
+
+          val destination = if (to is FileSystemDirectory) {
+            val contentDisposition = headerFields["Content-Disposition"]?.first()
+            val fileName = URLUtil.guessFileName(url.toString(), contentDisposition, contentType)
+            File(to.path, fileName)
+          } else {
+            to.path
+          }
+          FileOutputStream(destination).use { output ->
+            inputStream.copyTo(output)
+            promise.resolve(destination.path)
+          }
+        }
+      } catch (e: Exception) {
+        promise.reject(UnableToDownloadException(e.message))
+        return@AsyncFunction
+      }
+    }
 
     Class(FileSystemFile::class) {
       Constructor { path: URI ->
@@ -52,8 +82,16 @@ class FileSystemNextModule : Module() {
         file.exists()
       }
 
+      Function("copy") { file: FileSystemFile, destination: FileSystemPath ->
+        file.copy(destination)
+      }
+
+      Function("move") { file: FileSystemFile, destination: FileSystemPath ->
+        file.move(destination)
+      }
+
       Property("path")
-        .get { file: FileSystemFile -> return@get Uri.fromFile(file.path) }
+        .get { file: FileSystemFile -> return@get file.asString() }
     }
 
     Class(FileSystemDirectory::class) {
@@ -77,8 +115,16 @@ class FileSystemNextModule : Module() {
         directory.validatePath()
       }
 
+      Function("copy") { directory: FileSystemDirectory, destination: FileSystemPath ->
+        directory.copy(destination)
+      }
+
+      Function("move") { directory: FileSystemDirectory, destination: FileSystemPath ->
+        directory.move(destination)
+      }
+
       Property("path")
-        .get { directory: FileSystemDirectory -> return@get Uri.fromFile(directory.path) }
+        .get { directory: FileSystemDirectory -> return@get directory.asString() }
     }
   }
 }