+++ packages/expo-file-system/ios/Next/FileSystemFile.swift
@@ -2,9 +2,24 @@ import Foundation
 import ExpoModulesCore
 
 internal final class FileSystemFile: FileSystemPath {
+  override init(url: URL) {
+    if !url.hasDirectoryPath {
+      super.init(url: url)
+      return
+    }
+    super.init(url: url.deletingLastPathComponent().appendingPathComponent(url.lastPathComponent, isDirectory: false))
+  }
   func create() {
     FileManager.default.createFile(atPath: url.path, contents: nil)
   }
+  func exists() -> Bool {
+    var isDirectory: ObjCBool = false
+    if FileManager.default.fileExists(atPath: url.path, isDirectory: &isDirectory) {
+      return !isDirectory.boolValue
+    }
+    return false
+  }
+  // TODO: Move to the constructor once error is rethrowed
   func validatePath() throws {
     guard url.isFileURL && !url.hasDirectoryPath else {
       throw Exception(name: "wrong type", description: "tried to create a file with a directory path")