+++ packages/expo-file-system/ios/Next/FileSystemPath.swift
@@ -11,8 +11,27 @@ internal class FileSystemPath: SharedObject {
   func delete() throws {
     try FileManager.default.removeItem(at: url)
   }
+  func copy(to: FileSystemPath) throws {
+    if to is FileSystemDirectory {
+      try FileManager.default.copyItem(at: url, to: to.url.appendingPathComponent(url.lastPathComponent))
+    }
+    if to is FileSystemFile {
+      guard !url.hasDirectoryPath else {
+        throw CopyFolderToFileException()
+      }
+      try FileManager.default.copyItem(at: url, to: to.url)
+    }
+  }
 
-  func exists() -> Bool {
-    return FileManager.default.fileExists(atPath: url.path)
+  func move(to: FileSystemPath) throws {
+    if to is FileSystemDirectory {
+      try FileManager.default.moveItem(at: url, to: to.url.appendingPathComponent(url.lastPathComponent))
+    }
+    if to is FileSystemFile {
+      guard !url.hasDirectoryPath else {
+        throw MoveFolderToFileException()
+      }
+      try FileManager.default.moveItem(at: url, to: to.url)
+    }
   }
 }